third_party/WebKit/LayoutTests/bluetooth/characteristic/getDescriptors/gen-characteristic-is-removed-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let service, characteristic, fake_service, fake_characteristic;
  return getMeasurementIntervalCharacteristic()
    .then(vals => ({
      service,
      characteristic,
      fake_service,
      fake_characteristic,
    } = vals))
    .then(() => characteristic.getDescriptor(user_description.name))
    .then(() => null, (e) => assert_unreached('Caught error unexpectedly.', e))
    .then(() => fake_service.removeFakeCharacteristic({
      characteristic: fake_characteristic,
    }))
    .then(() => characteristic.getDescriptors(user_description.name))
    .then(() => assert_unreached('Got descriptor unexpectedly.'), (e) => {
      assert_equals(e.name, 'InvalidStateError');
      assert_equals(e.message, 'GATT Characteristic no longer exists.');
    });
}, 'Characteristic gets removed. Reject with InvalidStateError.');

</script>
